monitor/requirements.txt
requests==2.32.3
boto3==1.34.84
pydantic-settings==2.2.1
